Refactor snapshot into own file
